Add optional basic auth
You now can set a password in the SHELL_CONFIG variable. A basic auth
prompt will ask for the password when opening the page (user name is
ignored). When authpass is set to a blank string the feature is disabled
again, just like before.

This should give some peace of mind when using this script as it should
prevent 3rd parties to access it while you use it.